API updates for improved sensor data structure

+++ src/main/java/uk/ac/oak/movemore/webapp/model/ObsvOBDDetection.java
@@ -74,14 +74,14 @@ public class ObsvOBDDetection extends BaseObject implements Serializable {
 	private Double intakeManifoldPressure;
 	
 	// observation location
-	private Float longitude;
-	private Float latitude;
+	private Double longitude;
+	private Double latitude;
 	
 	public ObsvOBDDetection() {
 		
 	}
 	
-	public ObsvOBDDetection (Float longitude, Float latitude, Date obsvTime) {
+	public ObsvOBDDetection (Double longitude, Double latitude, Date obsvTime) {
 		setLongitude(longitude);
 		setLatitude(latitude);
 		setObsvTime(obsvTime);
@@ -171,20 +171,20 @@ public class ObsvOBDDetection extends BaseObject implements Serializable {
 	}
 
 	@Column(name = "longitude", precision = 13, scale = 10, updatable = false)
-	public Float getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Float longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
 
 	@Column(name = "latitude", precision = 13, scale = 10, updatable = false)
-	public Float getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Float latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
