+++ src/main/java/uk/ac/oak/movemore/webapp/model/ObsvPeopleCount.java
@@ -36,8 +36,8 @@ public class ObsvPeopleCount extends BaseObject implements Serializable {
 	private Date obsvTime;
 
 	// observation location
-	private Float longitude;
-	private Float latitude;
+	private Double longitude;
+	private Double latitude;
 
 	public ObsvPeopleCount() {
 	}
@@ -97,20 +97,20 @@ public class ObsvPeopleCount extends BaseObject implements Serializable {
 	}
 
 	@Column(name = "longitude", precision = 13, scale = 10, updatable=false)
-	public Float getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Float longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
 
 	@Column(name = "latitude", precision = 13, scale = 10, updatable=false)
-	public Float getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Float latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
