+++ src/main/java/uk/ac/oak/movemore/webapp/service/response/ObservationDetail.java
@@ -21,8 +21,8 @@ public class ObservationDetail extends SensorObservationSuccess implements
 
 	private String sensorName;
 
-	private Float longitude;
-	private Float latitude;
+	private Double longitude;
+	private Double latitude;
 
 	public Date getObsvTime() {
 		return obsvTime;
@@ -64,19 +64,19 @@ public class ObservationDetail extends SensorObservationSuccess implements
 		this.sensorName = sensorName;
 	}
 
-	public Float getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Float longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
 
-	public Float getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Float latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
