+++ src/main/java/uk/ac/oak/movemore/webapp/vo/DeviceVO.java
@@ -14,8 +14,8 @@ public class DeviceVO implements Serializable{
 	private String name;
 	private String description;
 	private String devicePhysicalId;
-	private Float longitude;
-	private Float latitude;
+	private Double longitude;
+	private Double latitude;
 	private String lastKnownIP;
 	private Float batteryLevel;
 	
@@ -43,16 +43,16 @@ public class DeviceVO implements Serializable{
 	public void setDevicePhysicalId(String devicePhysicalId) {
 		this.devicePhysicalId = devicePhysicalId;
 	}
-	public Float getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(Float longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
-	public Float getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(Float latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 	public String getLastKnownIP() {
