+++ src/test/java/uk/ac/oak/movemore/webapp/dao/ObsvOBDSensorDaoTest.java
@@ -35,8 +35,8 @@ public class ObsvOBDSensorDaoTest extends BaseDaoTestCase {
 		Double ambientAirTemperature=25.0;
 		Double engineRPM = 899.0;
 		Double fuelLevel= 0.0;
-		Float latitude = 53.383f;
-		Float longitude = -1.4659f;
+		Double latitude = 53.383d;
+		Double longitude = -1.4659d;
 		Date obsvTime = new Date();
 		
 		Sensors dummySensor = sensorsDao.get(sensorId);
